requirements.txt
# EngageWatch - Dependencies
# Python 3.10+ required

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Browser automation
playwright>=1.41.0

# Database
aiosqlite>=0.19.0

# Notifications
python-telegram-bot>=20.7
aiosmtplib>=3.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP client (for Telegram)
httpx>=0.26.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-timeout>=2.2.0
respx>=0.20.0
